Split device settings and xochitlsettings into their own files
