+++ shared/liboxide/liboxide.h
@@ -18,6 +18,8 @@
 #include "slothandler.h"
 #include "sysobject.h"
 #include "debug.h"
+#include "devicesettings.h"
+#include "xochitlsettings.h"
 #if defined(LIBOXIDE_LIBRARY)
 #include "oxide_sentry.h"
 #else
@@ -45,26 +47,11 @@ Q_IMPORT_PLUGIN(QsgEpaperPlugin)
 #define APP_VERSION VERSION
 #endif
 /*!
- * \def deviceSettings()
- * \brief Get the Oxide::DeviceSettings instance
- */
-#define deviceSettings Oxide::DeviceSettings::instance()
-/*!
- * \def xochitlSettings()
- * \brief Get the Oxide::XochitlSettings instance
- */
-#define xochitlSettings Oxide::XochitlSettings::instance()
-/*!
  * \def sharedSettings()
  * \brief Get the Oxide::SharedSettings instance
  */
 #define sharedSettings Oxide::SharedSettings::instance()
 
-/*!
- * \brief Wifi Network definition
- */
-typedef QMap<QString, QVariantMap> WifiNetworks;
-Q_DECLARE_METATYPE(WifiNetworks);
 /*!
  * \brief The main Oxide namespace
  */
@@ -148,204 +135,7 @@ namespace Oxide {
      * \snippet examples/oxide.cpp getGID
      */
     LIBOXIDE_EXPORT gid_t getGID(const QString& name);
-    /*!
-     * \brief Device specific values
-     */
-    class LIBOXIDE_EXPORT DeviceSettings{
-    public:
-        /*!
-         * \brief Known device types
-         */
-        enum DeviceType {
-            Unknown, /*!< Unknown device type >*/
-            RM1, /*!< reMarkable 1 >*/
-            RM2 /*!< reMarkable 2 >*/
-        };
-        /*!
-         * \brief Get the static instance of this class. You should use the deviceSettings macro instead.
-         * \return The static instance
-         * \sa deviceSettings
-         */
-        static DeviceSettings& instance();
-        /*!
-         * \brief Get the path to the buttons input device
-         * \return Path to the buttons device
-         */
-        const char* getButtonsDevicePath() const;
-        /*!
-         * \brief Get the path to the wacom input device
-         * \return Path to the wacom device
-         */
-        const char* getWacomDevicePath() const;
-        /*!
-         * \brief Get the path to the touch input device
-         * \return Path to the touch device
-         */
-        const char* getTouchDevicePath() const;
-        /*!
-         * \brief Get the Qt environment settings for the device
-         * \return The Qt environment settings for the device
-         */
-        const char* getTouchEnvSetting() const;
-        /*!
-         * \brief Get the device type
-         * \return The device type
-         */
-        DeviceType getDeviceType() const;
-        /*!
-         * \brief Get the human readable device name
-         * \return Human readable device name
-         */
-        const char* getDeviceName() const;
-        /*!
-         * \brief Get the max width for touch input on the device
-         * \return Max width for touch input
-         */
-        int getTouchWidth() const;
-        /*!
-         * \brief Get the max height for touch input on the device
-         * \return Max height for touch input
-         */
-        int getTouchHeight() const;
-        /*!
-         * \brief Get the list of possible locales on the device
-         * \return The list of possible locales on the device
-         */
-        const QStringList getLocales();
-        /*!
-         * \brief Get the current set locale
-         * \return The current locale
-         */
-        QString getLocale();
-        /*!
-         * \brief Set the current locale
-         * \param locale Locale to set
-         */
-        void setLocale(const QString& locale);
-        /*!
-         * \brief Get the list of possible timezones on the device
-         * \return The list of possible timezones on the device
-         */
-        const QStringList getTimezones();
-        /*!
-         * \brief Get the current set timezone
-         * \return The current timezone
-         */
-        QString getTimezone();
-        /*!
-         * \brief Set the current timezone
-         * \param locale Timezone to set
-         */
-        void setTimezone(const QString& timezone);
-        /*!
-         * \brief Setup the Qt environment
-         * \snippet examples/oxide.cpp setupQtEnvironment
-         */
-        void setupQtEnvironment(bool touch = true);
-        /*!
-         * \brief Check if a keyboard is attached
-         * \return If a keyboard is attached
-         */
-        bool keyboardAttached();
-        /*!
-         * \brief Run a callback when keyboardAttached changes
-         * \param callback Callback to run
-         */
-        void onKeyboardAttachedChanged(std::function<void()> callback);
 
-    private:
-        DeviceType _deviceType;
-
-        DeviceSettings();
-        ~DeviceSettings();
-        void readDeviceType();
-        bool checkBitSet(int fd, int type, int i);
-        std::string buttonsPath = "";
-        std::string wacomPath = "";
-        std::string touchPath = "";
-    };
-    /*!
-     * \brief Manage Xochitl settings
-     */
-    class LIBOXIDE_EXPORT XochitlSettings : public SettingsFile {
-        Q_OBJECT
-        /*!
-         * \fn instance
-         * \brief Get the static instance of this class. You should use the xochitlSettings macro instead.
-         * \return The static instance
-         * \sa xochitlSettings
-         */
-        // cppcheck-suppress uninitMemberVarPrivate
-        // cppcheck-suppress unusedFunction
-        O_SETTINGS(XochitlSettings, "/home/root/.config/remarkable/xochitl.conf")
-        /*!
-         * \property passcode
-         * \brief The passcode used to unlock the device
-         * \sa set_passcode, passcodeChanged
-         */
-        /*!
-         * \fn set_passcode
-         * \param _arg_passcode The passcode used to unlock the device
-         * \brief Set the passcode used to unlock the device
-         */
-        /*!
-         * \fn passcodeChanged
-         * \brief The passcode used to unlock the device has changed
-         */
-        O_SETTINGS_PROPERTY(QString, General, passcode)
-        /*!
-         * \property wifion
-         * \brief If wifi is on or off
-         * \sa set_wifion, wifionChanged
-         */
-        /*!
-         * \fn set_wifion
-         * \param _arg_wifion If wifi should be on or off
-         * \brief Turn wifi on or off
-         */
-        /*!
-         * \fn wifionChanged
-         * \brief Wifi has been turned on or off
-         */
-        O_SETTINGS_PROPERTY(bool, General, wifion)
-        /*!
-         * \property XochitlSettings::wifinetworks
-         * \brief List of wifi networks
-         * \sa setWifinetworks, wifinetworksChanged
-         */
-        Q_PROPERTY(WifiNetworks wifinetworks MEMBER m_wifinetworks READ wifinetworks WRITE setWifinetworks RESET resetWifinetworks NOTIFY wifinetworksChanged)
-
-    public:
-        WifiNetworks wifinetworks();
-        /*!
-         * \brief Set the list of wifi networks
-         * \param wifinetworks List of wifi networks to replace with
-         */
-        void setWifinetworks(const WifiNetworks& wifinetworks);
-        /*!
-         * \brief Get a specific wifi network
-         * \param name SSID of the wifi network
-         * \return The wifi network properties
-         */
-        QVariantMap getWifiNetwork(const QString& name);
-        /*!
-         * \brief Set the properties for a specific wifi network
-         * \param name SSID of the wifi network
-         * \param properties The wifi network properties
-         */
-        void setWifiNetwork(const QString& name, QVariantMap properties);
-        void resetWifinetworks();
-
-    signals:
-        /*!
-         * \brief The contents of the wifi network list has changed
-         */
-        void wifinetworksChanged(WifiNetworks);
-
-    private:
-        ~XochitlSettings();
-        WifiNetworks m_wifinetworks;
-    };
     /*!
      * \brief Shared settings for Oxide
      */
